+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added support for setting the kernel's backlog wait time via the new
+  SetBacklogWaitTime function. #34
+
 ### Changed
 
 ### Deprecated